backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy
passlib[bcrypt]
bcrypt
psycopg2-binary
python-dotenv